requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Django==3.1
django-cors-headers==3.2.0
idna==2.10
mysql-connector-python==8.0.18
mysqlclient==1.4.6
protobuf==3.11.2
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
requests==2.24.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.10
